Add pibridge simulation programs
Add simulate_pibridge_master and simulate_pibridge_modules which
communicate with each other and simulate the pibrige communication over the
uart. The serial line communication settings have to be set up first, e.g.
by means of 'stty'.

The communication shows the same type of errors that can be observed during
the communication of a RevPi master and RevPi modules on the real
pibridge.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>